+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClientFactory.java
@@ -1,24 +1,21 @@
 package com.getindata.connectors.http.internal.table.lookup;
 
 import java.net.http.HttpClient;
-import java.net.http.HttpClient.Redirect;
 
 import org.apache.flink.api.common.serialization.DeserializationSchema;
 import org.apache.flink.table.data.RowData;
 
-import com.getindata.connectors.http.internal.PollingClient;
 import com.getindata.connectors.http.internal.PollingClientFactory;
+import com.getindata.connectors.http.internal.utils.JavaNetHttpClientFactory;
 
 public class JavaNetHttpPollingClientFactory implements PollingClientFactory<RowData> {
 
     @Override
-    public PollingClient<RowData> createPollClient(
+    public JavaNetHttpPollingClient createPollClient(
             HttpLookupConfig options,
             DeserializationSchema<RowData> schemaDecoder) {
 
-        HttpClient httpClient = HttpClient.newBuilder()
-            .followRedirects(Redirect.NORMAL)
-            .build();
+        HttpClient httpClient = JavaNetHttpClientFactory.createClient(options.getProperties());
 
         return new JavaNetHttpPollingClient(httpClient, schemaDecoder, options);
     }