+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add TLS and mTLS support for Http Sink and Lookup Source connectors.  
+New properties are:
+  - `gid.connector.http.security.cert.server` - path to server's certificate.
+  - `gid.connector.http.security.cert.client` - path to connector's certificate.
+  - `gid.connector.http.security.key.client` - path to connector's private key.
+  - `gid.connector.http.security.cert.server.allowSelfSigned` - allowing for self signed certificates without adding them to KeyStore (not recommended for a production).
+
 ## [0.4.0] - 2022-08-31
 
 ### Added